--- src/OrdersList/OrdersList.cpp
@@ -1,14 +1,0 @@
-#include<iostream>
-using namespace std;
-
-#include "OrdersList.h"
-
-// Define the order class here
-
- // move()
-
- // remove()
-
- // validate()
-
- // execute()
